feat(database): refactor database initialization and encryption validation; remove vector extension handling
