+++ .env.example
@@ -7,4 +7,5 @@ FEEDBACK_WEBHOOK=https://discord.com/api/webhooks/your_webhook_id/your_webhook_t
 POSTGRES_USER=saltbot
 POSTGRES_PASSWORD=strong_password_here
 POSTGRES_DB=saltbot
-POSTGRES_URI=postgres://${POSTGRES_USER}:${POSTGRES_PASSWORD}@postgres:5432/${POSTGRES_DB}
+## Note: Do not use docker postgresdb in production, use a managed database service instead.like neondb https://console.neon.tech/
+POSTGRES_URI=postgres://${POSTGRES_USER}:${POSTGRES_PASSWORD}@postgres:5432/${POSTGRES_DB}
